split crypto module
